bem, move styles, components tree
