+++ src/app/transactions/transactions.component.html
@@ -1,3 +1,5 @@
-<button mat-fab (click)="handleAddTransactionClick()" class="add-transaction-dialog-activation">
-  <mat-icon aria-label="add">add</mat-icon>
-</button>
+<div class="transactions">
+  <button mat-fab (click)="handleAddTransactionClick()" class="transactions__add-dialog-activation">
+    <mat-icon aria-label="add">add</mat-icon>
+  </button>
+</div>
