+++ src/app/transactions/transactions.component.scss
@@ -1,5 +1,7 @@
-.add-transaction-dialog-activation {
-  position: fixed;
-  right: 12px;
-  bottom: 12px;
+.transactions {
+  &__add-dialog-activation {
+    position: fixed;
+    right: 12px;
+    bottom: 12px;
+  }
 }
